users can only edit and delete their own tasks

+++ src/Controller/TasksController.php
@@ -30,6 +30,7 @@ class TasksController extends AppController
         if ($this->request->is('post')){
             $task = $this->Tasks->patchEntity($task, $this->request->getData());
             $task->Owner=$this->Auth->user('name');
+            $task->user_id=$this->Auth->user('id');
             if($this->Tasks->save($task)){
                 $this->Flash->success("Deine Aufgabe wurde notiert!");
                 return $this->redirect(['action'=>'index']);
@@ -49,9 +50,16 @@ class TasksController extends AppController
 
     public function edit($TaskId = null)
     {
+
         $task = $this->Tasks->get($TaskId, [
             'contain' => []
         ]);
+       if( $task->user_id!==$this->Auth->user('id')) {
+            $this->Flash->error('Du kannst nur deine eigenen Aufgaben bearbeiten!');
+            return $this->redirect(['action' => 'index']);
+       }
+       
+
         if ($this->request->is(['patch', 'post', 'put'])) {
             $task = $this->Tasks->patchEntity($task, $this->request->getData());
             if ($this->Tasks->save($task)) {
@@ -66,8 +74,15 @@ class TasksController extends AppController
        
     public function delete($TaskId = null)
     {
-        $this->request->allowMethod(['post', 'delete']);
+
         $task = $this->Tasks->get($TaskId);
+
+        if( $task->user_id!==$this->Auth->user('id')) {
+            $this->Flash->error('Du kannst nur deine eigenen Aufgaben löschen!');
+            return $this->redirect(['action' => 'index']);
+       }
+
+        $this->request->allowMethod(['post', 'delete']);
         if ($this->Tasks->delete($task)) {
             $this->Flash->success(__('Die Aufgabe wurde gelöscht.'));
         } else {
